cs-validator: move package from mapexplorer-core
