+++ packages/cs-validator/.npmignore
@@ -1,0 +1,9 @@
+.DS_Store
+*.log
+*.pdf
+test
+coverage
+_book
+dist
+.idea
+
