+++ packages/cs-validator/rollup.browser.config.js
@@ -1,0 +1,12 @@
+const plugins = require('./rollupPlugins');
+
+export default {
+  plugins,
+  sourcemap: true,
+  input: 'src/index.js',
+  output: {
+    file: 'dist/stratumn-cs-validator.js',
+    format: 'umd'
+  },
+  name: 'stratumnCsValidator'
+};
