+++ packages/cs-validator/src/MapValidator.js
@@ -14,12 +14,12 @@
   limitations under the License.
 */
 
+import { resolveLinks } from 'stratumn-agent-client';
 import SegmentValidator from './SegmentValidator';
-import resolveLinks from './resolveLinks';
 import wrap from './wrap';
 import parseIfJson from './parseIfJson';
 
-export default class ChainValidator {
+export default class MapValidator {
   constructor(chainscript) {
     this.chainscript = chainscript;
     this.errors = {
