+++ packages/cs-validator/src/computeMerkleParent.js
