+++ packages/cs-validator/src/hashJson.js
