+++ packages/cs-validator/test/fixtures/fullMap.json
@@ -1,0 +1,194 @@
+[
+  {
+    "link": {
+      "state": {
+        "gameId": "Point"
+      },
+      "meta": {
+        "mapId": "5770eeadd6a22a8f2bd2bbf6",
+        "agentHash":
+          "e0b0976355e77077905db39c8dee0879b82e24f962853fa50c87799f8e33d19f",
+        "stateHash":
+          "2dffd2e02f4ae7daf2bc0bb61cada10b1ee023b1bc1be27474369fd67cd50113",
+        "prevLinkHash": null,
+        "action": "init",
+        "arguments": ["Point"]
+      }
+    },
+    "meta": {
+      "linkHash":
+        "32fc2804fb82472a44cafdd325745706318d9bed958fef14b535d91fb9527287",
+      "evidences": [
+        {
+          "state": "COMPLETE",
+          "backend": "bcbatch",
+          "provider": "bcbatch",
+          "proof": {
+            "batch": {
+              "merkleRoot":
+                "35b8052870a97136a2382b7d7d5528feb3372be1b0ab3414eeb917c2245885d9",
+              "merklePath": [
+                {
+                  "left":
+                    "32fc2804fb82472a44cafdd325745706318d9bed958fef14b535d91fb9527287",
+                  "right":
+                    "34cd7b3fec935f1037b8c8308db09b70255caa7db56c7cfb9b7e34e19682c0ca",
+                  "parent":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff"
+                },
+                {
+                  "left":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff",
+                  "parent":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff"
+                },
+                {
+                  "left":
+                    "b337d23025b9c5a15b5645a9473c3b3ff904b9c1a25d7fb617c19944082ca6f8",
+                  "right":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff",
+                  "parent":
+                    "18d5ff8abb9c02107b1f54096563b46f37d0a3b9acb93fb82c6eee2a99fb4013"
+                },
+                {
+                  "left":
+                    "a6893894a1f37d9e062e1fb26322a59dcac2c75d1096ac84fd95c1976d13bc64",
+                  "right":
+                    "18d5ff8abb9c02107b1f54096563b46f37d0a3b9acb93fb82c6eee2a99fb4013",
+                  "parent":
+                    "35b8052870a97136a2382b7d7d5528feb3372be1b0ab3414eeb917c2245885d9"
+                }
+              ]
+            },
+            "txid":
+              "2a443211e871f58a6ee5a93e62ce36cac2ddfc0f05a6bec1e7b11aa8d5e4cf38"
+          }
+        }
+      ],
+      "application": "snake-eyes",
+      "applicationLocation": "https://snake-eyes.stratumn.net",
+      "linkLocation":
+        "https://snake-eyes.stratumn.net/links/32fc2804fb82472a44cafdd325745706318d9bed958fef14b535d91fb9527287",
+      "xStratumn": {
+        "totalTimeMs": 7.225957999999999,
+        "loadApplicationCache": "hit",
+        "loadApplicationTimeMs": 0.030917999999999998,
+        "loadOwnerIdCache": "hit",
+        "loadOwnerIdTimeMs": 0.012791,
+        "loadInputLinkCache": "miss",
+        "loadInputLinkTimeMs": 7.082053999999999
+      }
+    }
+  },
+  {
+    "link": {
+      "state": {
+        "gameId": "Point",
+        "nick": "Adrien",
+        "address": "1EQW6dpL4YNicmpZSBZ8KKzz1aeGA5GaYZ",
+        "messages": [],
+        "rolls": 0
+      },
+      "meta": {
+        "mapId": "5770eeadd6a22a8f2bd2bbf6",
+        "agentHash":
+          "e0b0976355e77077905db39c8dee0879b82e24f962853fa50c87799f8e33d19f",
+        "stateHash":
+          "6c20f8b575d0c6a82611b63342f6ef4e9cdfed4956b3e81bd6ba462d4d2b6231",
+        "prevLinkHash":
+          "32fc2804fb82472a44cafdd325745706318d9bed958fef14b535d91fb9527287",
+        "action": "register",
+        "arguments": ["Adrien", "1EQW6dpL4YNicmpZSBZ8KKzz1aeGA5GaYZ"],
+        "tags": ["win"]
+      }
+    },
+    "meta": {
+      "linkHash":
+        "dfc29526002ff024c7caab3cc462c6d5e227eea8e8827e49113333cd679045d1",
+      "evidences": [
+        {
+          "state": "COMPLETE",
+          "backend": "bcbatch",
+          "provider": "bcbatch",
+          "proof": {
+            "batch": {
+              "merkleRoot":
+                "35b8052870a97136a2382b7d7d5528feb3372be1b0ab3414eeb917c2245885d9",
+              "merklePath": [
+                {
+                  "left":
+                    "dfc29526002ff024c7caab3cc462c6d5e227eea8e8827e49113333cd679045d1",
+                  "right":
+                    "aa5b42183dc572274ae67d7c2eb106f912673ce8a7d613b300921177d55ee8d5",
+                  "parent":
+                    "138224a59b9091b54a70f479a190f7a3fea7018851c6244e1bcffbe77f83e7fc"
+                },
+                {
+                  "left":
+                    "138224a59b9091b54a70f479a190f7a3fea7018851c6244e1bcffbe77f83e7fc",
+                  "right":
+                    "6a2c78b29323243572ac2a68269c91374b2406c670fe5acb3c6848f50e51a22b",
+                  "parent":
+                    "5fc38c900554f137528f21a6371639e242fe1f633865d6c67b89b434862891ad"
+                },
+                {
+                  "left":
+                    "5fc38c900554f137528f21a6371639e242fe1f633865d6c67b89b434862891ad",
+                  "right":
+                    "5b97214099f6a791940fb16acacde29b3d41301bc35484012171b69b73532d11",
+                  "parent":
+                    "34cd7b3fec935f1037b8c8308db09b70255caa7db56c7cfb9b7e34e19682c0ca"
+                },
+                {
+                  "left":
+                    "32fc2804fb82472a44cafdd325745706318d9bed958fef14b535d91fb9527287",
+                  "right":
+                    "34cd7b3fec935f1037b8c8308db09b70255caa7db56c7cfb9b7e34e19682c0ca",
+                  "parent":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff"
+                },
+                {
+                  "left":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff",
+                  "parent":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff"
+                },
+                {
+                  "left":
+                    "b337d23025b9c5a15b5645a9473c3b3ff904b9c1a25d7fb617c19944082ca6f8",
+                  "right":
+                    "3333e49db57ec5ea1ec37c21ad9d5de4f10c550fe858f88f66feaa8e01e85bff",
+                  "parent":
+                    "18d5ff8abb9c02107b1f54096563b46f37d0a3b9acb93fb82c6eee2a99fb4013"
+                },
+                {
+                  "left":
+                    "a6893894a1f37d9e062e1fb26322a59dcac2c75d1096ac84fd95c1976d13bc64",
+                  "right":
+                    "18d5ff8abb9c02107b1f54096563b46f37d0a3b9acb93fb82c6eee2a99fb4013",
+                  "parent":
+                    "35b8052870a97136a2382b7d7d5528feb3372be1b0ab3414eeb917c2245885d9"
+                }
+              ]
+            },
+            "txid":
+              "2a443211e871f58a6ee5a93e62ce36cac2ddfc0f05a6bec1e7b11aa8d5e4cf38"
+          }
+        }
+      ],
+      "application": "snake-eyes",
+      "applicationLocation": "https://snake-eyes.stratumn.net",
+      "linkLocation":
+        "https://snake-eyes.stratumn.net/links/dfc29526002ff024c7caab3cc462c6d5e227eea8e8827e49113333cd679045d1",
+      "xStratumn": {
+        "totalTimeMs": 6.31556,
+        "loadApplicationCache": "hit",
+        "loadApplicationTimeMs": 0.020272,
+        "loadOwnerIdCache": "hit",
+        "loadOwnerIdTimeMs": 0.009373,
+        "loadInputLinkCache": "miss",
+        "loadInputLinkTimeMs": 6.181551
+      }
+    }
+  }
+]
