+++ packages/mapexplorer-core/README.md
@@ -159,32 +159,6 @@ where `merklePath` looks like:
 ]
 ```
 
-### Validate a chainscript
-
-```javascript
-new MapexplorerCore.ChainValidator(JSON.parse(chainscript)).validate()
-```
-
-Returns a Promise that resolves to an error object such as:
-
-```
-{
-      linkHash: [Promise, ...],
-      stateHash: [Promise, ...],
-      merklePath: [Promise, ...],
-      fossil: [Promise, ...]
-}
-```
-
-Each promise, in each array of each category resolves to an error string if an inconsistency has been detected. It resolves to null otherwise.
-
-Errors can be retrieved with (for instance):
-
-```javascript
-Promise.all(errors.linkHash).
-  then(err => err.filter(Boolean));
-```
-
 ## Development
 
 Install dependencies:
