+++ packages/mapexplorer-core/src/ChainTreeBuilder.js
@@ -14,13 +14,12 @@
   limitations under the License.
 */
 
-import { getAgent } from 'stratumn-agent-client';
+import { getAgent, resolveLinks } from 'stratumn-agent-client';
 import ChainTree from './ChainTree';
 import compactHash from './compactHash';
-import resolveLinks from './resolveLinks';
+import tagsSet from './tagsSet';
 import wrap from './wrap';
 import parseIfJson from './parseIfJson';
-import tagsSet from './tagsSet';
 
 function load(map) {
   return getAgent(map.agentUrl)
